refactor: separate recipe details part to components
