+++ components/submissions/RecipeLayout.tsx
@@ -1,0 +1,136 @@
+import {
+    Ingredient,
+    Instruction,
+    Recipe,
+} from '@/pages/submissions/types/submission'
+import moment from 'moment'
+import { Fragment, useMemo } from 'react'
+import Image from 'next/image'
+
+interface RecipeLayoutProps {
+    content: Recipe
+}
+
+interface IngredientsProps {
+    ingredients: Ingredient[]
+}
+
+interface InstructionProps {
+    instructions: Instruction[]
+}
+
+export const RecipeInfoCard = ({ content }: RecipeLayoutProps) => {
+    const timeSpanToString = (timespan: string): string => {
+        const duration = moment.duration(timespan)
+
+        let result = ''
+        if (duration.days() !== 0) {
+            result += duration.days() + ' days '
+        }
+        if (duration.minutes() !== 0) {
+            result += duration.minutes() + ' minutes '
+        }
+        if (duration.seconds() !== 0) {
+            result += duration.seconds() + ' seconds'
+        }
+
+        return result.trim()
+    }
+
+    const prepTimeString = useMemo(
+        () => timeSpanToString(content.prepTime),
+        [content]
+    )
+    const cookTimeString = useMemo(
+        () => timeSpanToString(content.cookTime),
+        [content]
+    )
+
+    return (
+        <div className="grid grid-cols-4 px-4 py-4 rounded-lg bg-cprimary-300 text-white">
+            <div className="flex flex-col items-center justify-center">
+                <h2 className="text-xl">Difficulty</h2>
+                <p>{content.difficulty}</p>
+            </div>
+            <div className="flex flex-col items-center justify-center">
+                <h2 className="text-xl">Portions</h2>
+                <p>
+                    {content.portionQuantity} {content.portionType}
+                </p>
+            </div>
+            <div className="flex flex-col items-center justify-center">
+                <h2 className="text-xl">Prep time</h2>
+                <p>{prepTimeString}</p>
+            </div>
+            <div className="flex flex-col items-center justify-center">
+                <h2 className="text-xl">Cook time</h2>
+                <p>{cookTimeString}</p>
+            </div>
+        </div>
+    )
+}
+
+const RecipeIngredients = ({ ingredients }: IngredientsProps) => {
+    return (
+        <>
+            <div className="text-2xl">Ingredients</div>
+            <div className="grid grid-cols-3 mt-4">
+                {ingredients.map((ingredient, index) => (
+                    <Fragment key={index}>
+                        <div className="col-span-1 mb-2">
+                            {ingredient.amount} {ingredient.unit}
+                        </div>
+                        <div className="col-span-2 mb-2">{ingredient.name}</div>
+                    </Fragment>
+                ))}
+            </div>
+        </>
+    )
+}
+
+const RecipeInstructions = ({ instructions }: InstructionProps) => {
+    return (
+        <>
+            <div className="text-2xl">Instructions</div>
+            <div className="mt-4">
+                {instructions.map((instruction) => {
+                    const instructionCount = instructions.length
+                    return (
+                        <div key={instruction.step} className="mb-8">
+                            <div className="text-xl">
+                                Step {instruction.step}/{instructionCount}
+                            </div>
+                            <div className="mt-2">
+                                <Image
+                                    src={
+                                        'https://m.media-amazon.com/images/I/616U8Co1ASL.jpg'
+                                    }
+                                    width={1000}
+                                    height={1000}
+                                    style={{ objectFit: 'contain' }}
+                                    alt={`step ${instruction.step} illustration`}
+                                />
+                            </div>
+                            <p className="text-lg mt-2">
+                                {instruction.description}
+                            </p>
+                        </div>
+                    )
+                })}
+            </div>
+        </>
+    )
+}
+
+export const RecipeGuideSection = ({ content }: RecipeLayoutProps) => {
+    return (
+        <div className="grid grid-cols-3">
+            <div className="col-span-1">
+                <RecipeIngredients ingredients={content.ingredients} />
+            </div>
+            <div className="col-span-2">
+                <RecipeInstructions instructions={content.instructions} />
+            </div>
+        </div>
+    )
+}
